+++ app/src/main/java/org/ole/planet/myplanet/ui/myhealth/AddMyHealthActivity.java
@@ -19,6 +19,7 @@ import org.ole.planet.myplanet.utilities.AndroidDecrypter;
 import org.ole.planet.myplanet.utilities.Utilities;
 
 import java.util.List;
+import java.util.UUID;
 
 import io.realm.Realm;
 
@@ -42,14 +43,13 @@ public class AddMyHealthActivity extends AppCompatActivity {
         userId = getIntent().getStringExtra("userId");
         healthPojo = realm.where(RealmMyHealthPojo.class).equalTo("userId", userId).findFirst();
         userModelB = realm.where(RealmUserModel.class).equalTo("id", userId).findFirst();
-
         key = userModelB.getKey();
         iv = userModelB.getIv();
-        if (TextUtils.isEmpty(key) || TextUtils.isEmpty(iv)) {
-            Utilities.toast(this, "You cannot create health record from myPlanet. Please contact your manager.");
-            finish();
-            return;
-        }
+//        if (TextUtils.isEmpty(key) || TextUtils.isEmpty(iv)) {
+//            Utilities.toast(this, "You cannot create health record from myPlanet. Please contact your manager.");
+//            finish();
+//            return;
+//        }
         initViews();
         findViewById(R.id.btn_submit).setOnClickListener(view -> {
             createMyHealth();
@@ -76,10 +76,11 @@ public class AddMyHealthActivity extends AppCompatActivity {
             myHealth = new RealmMyHealth();
         myHealth.setProfile(health);
         if (healthPojo == null) {
-            healthPojo = realm.createObject(RealmMyHealthPojo.class, userId);
+            healthPojo = realm.createObject(RealmMyHealthPojo.class, UUID.randomUUID());
         }
         try {
-            healthPojo.setData(AndroidDecrypter.encrypt(new Gson().toJson(myHealth), key, iv));
+            healthPojo.setUserId(userId);
+            healthPojo.setData(TextUtils.isEmpty(userModelB.getIv()) ? new Gson().toJson(myHealth) : AndroidDecrypter.encrypt(new Gson().toJson(myHealth), key, iv));
         } catch (Exception e) {
         }
         finish();
